TD-875: Add plurals for new languages
